+++ lib/ratchetio/configuration.rb
@@ -8,6 +8,8 @@ module Ratchetio
     attr_accessor :default_logger
     attr_accessor :enabled
     attr_accessor :endpoint
+    attr_accessor :write_to_file
+    attr_accessor :filepath
     attr_accessor :environment
     attr_accessor :exception_level_filters
     attr_accessor :framework
@@ -18,6 +20,8 @@ module Ratchetio
     attr_accessor :person_email_method
     attr_accessor :root
     attr_accessor :scrub_fields
+    attr_accessor :use_async
+    attr_accessor :async_handler
 
     DEFAULT_ENDPOINT = 'https://submit.ratchet.io/api/1/item/'
 
@@ -25,6 +29,7 @@ module Ratchetio
       @default_logger = lambda { Logger.new(STDERR) }
       @enabled = true
       @endpoint = DEFAULT_ENDPOINT
+      @write_to_file = false
       @framework = 'Plain'
       @exception_level_filters = {
         'ActiveRecord::RecordNotFound' => 'warning',
@@ -36,6 +41,8 @@ module Ratchetio
       @person_username_method = 'username'
       @person_email_method = 'email'
       @scrub_fields = [:passwd, :password, :secret]
+      @use_async = false
+      @async_handler = nil
     end
 
     # allow params to be read like a hash